Crate tree2svg

Structs

Noeud
Un noeud présente un nom, un identifiant et une liste de descendants qui peut être vide. La liste de descendants est une référence car on veut garder la propriété (borower) Elle a le même temps de vie (lifetime) que le noeud père, ainsi que les noeuds de la liste. Elle est mutable car on peut ajouter des descendants.

Functions

explore_arbre_ref

main